+++ components/camel-infinispan/camel-infinispan/src/test/java/org/apache/camel/component/infinispan/remote/InfinispanRemoteConsumerIT.java
@@ -16,28 +16,33 @@
  */
 package org.apache.camel.component.infinispan.remote;
 
-import java.util.function.Supplier;
-
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.infinispan.InfinispanIdempotentRepositoryTestSupport;
+import org.apache.camel.component.infinispan.InfinispanConstants;
+import org.apache.camel.component.infinispan.InfinispanConsumerTestSupport;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.spi.IdempotentRepository;
-import org.apache.camel.util.function.Suppliers;
 import org.infinispan.commons.api.BasicCache;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledOnOs;
 import org.junit.jupiter.api.condition.OS;
 
 @DisabledOnOs(OS.MAC)
-public class InfinispanRemoteIdempotentRepositoryTest extends InfinispanRemoteTestSupport
-        implements InfinispanIdempotentRepositoryTestSupport {
+public class InfinispanRemoteConsumerIT extends InfinispanRemoteTestSupport implements InfinispanConsumerTestSupport {
+    @Test
+    public void consumerReceivedEventNotifications() throws Exception {
+        MockEndpoint mock = getMockEndpoint("mock:result");
+        mock.expectedMessageCount(1);
+        mock.expectedHeaderReceived(InfinispanConstants.EVENT_TYPE, "CLIENT_CACHE_ENTRY_CREATED");
 
-    private Supplier<IdempotentRepository> repo = Suppliers.memorize(() -> {
-        InfinispanRemoteIdempotentRepository repo = new InfinispanRemoteIdempotentRepository(getCacheName());
-        repo.setCacheContainer(cacheContainer);
+        getCache().put(InfinispanConsumerTestSupport.KEY_ONE, InfinispanConsumerTestSupport.VALUE_ONE);
+        getCache().put(InfinispanConsumerTestSupport.KEY_ONE, InfinispanConsumerTestSupport.VALUE_TWO);
 
-        return repo;
-    });
+        mock.assertIsSatisfied();
+    }
+
+    // *****************************
+    //
+    // *****************************
 
     @BeforeEach
     protected void beforeEach() {
@@ -46,18 +51,8 @@ public class InfinispanRemoteIdempotentRepositoryTest extends InfinispanRemoteTe
     }
 
     @Override
-    public IdempotentRepository getIdempotentRepository() {
-        return repo.get();
-    }
-
-    @Override
     public BasicCache<Object, Object> getCache() {
         return super.getCache();
-    }
-
-    @Override
-    public MockEndpoint getMockEndpoint(String id) {
-        return super.getMockEndpoint(id);
     }
 
     @Override
@@ -66,15 +61,16 @@ public class InfinispanRemoteIdempotentRepositoryTest extends InfinispanRemoteTe
     }
 
     @Override
+    public MockEndpoint getMockEndpoint(String id) {
+        return super.getMockEndpoint(id);
+    }
+
+    @Override
     protected RouteBuilder createRouteBuilder() throws Exception {
         return new RouteBuilder() {
             @Override
             public void configure() {
-                from("direct:start")
-                        .idempotentConsumer(
-                                header("MessageID"),
-                                getIdempotentRepository())
-                        .skipDuplicate(true)
+                fromF("infinispan:%s?eventTypes=CLIENT_CACHE_ENTRY_CREATED", getCacheName())
                         .to("mock:result");
             }
         };
