+++ components/camel-infinispan/camel-infinispan/src/test/java/org/apache/camel/component/infinispan/remote/cluster/AbstractInfinispanRemoteClusteredIT.java
@@ -53,7 +53,7 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @DisabledOnOs(OS.MAC)
-public class AbstractInfinispanRemoteClusteredTest {
+public class AbstractInfinispanRemoteClusteredIT {
     @RegisterExtension
     public static InfinispanService service = InfinispanServiceFactory.createService();
